+++ app/src/main/AndroidManifest.xml
@@ -7,12 +7,24 @@
 
     <application
             android:allowBackup="true"
+            android:fullBackupContent="@xml/backup_descriptor"
             android:icon="@mipmap/ic_launcher"
             android:label="@string/app_name"
             android:roundIcon="@mipmap/ic_launcher_round"
             android:supportsRtl="true"
-            android:theme="@style/AppTheme"
-            android:fullBackupContent="@xml/backup_descriptor">
+            android:theme="@style/AppTheme">
+        <activity android:name=".TvShowActivity"
+                  android:label="@string/title_activity_movie"
+                  android:parentActivityName=".MainActivity"
+                  android:theme="@style/AppTheme.DetailScreen">
+            <meta-data
+                    android:name="android.support.PARENT_ACTIVITY"
+                    android:value="com.kangmicin.hotmovie.MainActivity"/>
+
+            <intent-filter>
+                <action android:name="android.intent.action.VIEW"/>
+            </intent-filter>
+        </activity>
         <activity
                 android:name=".MovieActivity"
                 android:label="@string/title_activity_movie"
@@ -21,6 +33,7 @@
             <meta-data
                     android:name="android.support.PARENT_ACTIVITY"
                     android:value="com.kangmicin.hotmovie.MainActivity"/>
+
             <intent-filter>
                 <action android:name="android.intent.action.VIEW"/>
             </intent-filter>
@@ -32,7 +45,10 @@
                 <category android:name="android.intent.category.LAUNCHER"/>
             </intent-filter>
         </activity>
-        <meta-data android:name="preloaded_fonts" android:resource="@array/preloaded_fonts"/>
+
+        <meta-data
+                android:name="preloaded_fonts"
+                android:resource="@array/preloaded_fonts"/>
     </application>
 
 </manifest>
