+++ app/src/main/java/com/kangmicin/hotmovie/MainActivity.kt
@@ -3,6 +3,7 @@ package com.kangmicin.hotmovie
 import android.content.Intent
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
+import android.os.PersistableBundle
 import android.view.MenuItem
 import com.google.android.material.bottomnavigation.BottomNavigationView
 import com.kangmicin.hotmovie.model.Movie
@@ -13,9 +14,30 @@ class MainActivity : AppCompatActivity(), MovieContract.View, MovieListFragment.
 
     lateinit var presenter: MovieContract.Presenter
 
+    companion object {
+        var LAST_TAB = "LAST_TAB"
+    }
+
+    override fun onSaveInstanceState(outState: Bundle?, outPersistentState: PersistableBundle?) {
+        super.onSaveInstanceState(outState, outPersistentState)
+
+        outPersistentState?.run {
+            putInt(LAST_TAB, bottom_navigation.selectedItemId)
+        }
+    }
+
+    override fun onRestoreInstanceState(savedInstanceState: Bundle?) {
+        super.onRestoreInstanceState(savedInstanceState)
+
+        savedInstanceState?.run {
+            bottom_navigation.selectedItemId = getInt(LAST_TAB, bottom_navigation.selectedItemId)
+            bottom_navigation.dispatchSetSelected(false)
+        }
+    }
+
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
-
+        var initView = R.id.show_movie_menu
         val data = resources.obtainTypedArray(R.array.movies)
         val tvShowData = resources.obtainTypedArray(R.array.shows)
         val snapshot = Array<Array<String>>(data.length()) { i: Int ->
@@ -28,6 +50,10 @@ class MainActivity : AppCompatActivity(), MovieContract.View, MovieListFragment.
             resources.getStringArray(resourceId)
         }
 
+        savedInstanceState?.run {
+            initView = getInt(LAST_TAB, initView)
+        }
+
         presenter = MoviePresenter(this, {i, type ->
             when(type) {
                 MoviePresenter.ModelType.MOVIE -> snapshot[i]
@@ -37,6 +63,7 @@ class MainActivity : AppCompatActivity(), MovieContract.View, MovieListFragment.
 
         setContentView(R.layout.activity_main)
 
+        bottom_navigation?.selectedItemId = initView
         bottom_navigation?.setOnNavigationItemSelectedListener {
             when(it.itemId) {
                 R.id.show_movie_menu -> {
@@ -63,7 +90,6 @@ class MainActivity : AppCompatActivity(), MovieContract.View, MovieListFragment.
             .commit()
     }
 
-
     override fun displayTvShows(shows: List<TvShow>) {
         supportFragmentManager
             .beginTransaction()
@@ -79,6 +105,9 @@ class MainActivity : AppCompatActivity(), MovieContract.View, MovieListFragment.
     }
 
     override fun onListFragmentInteraction(item: TvShow?) {
+        val openIntent = Intent(this, TvShowActivity::class.java)
 
+        openIntent.putExtra(TvShowActivity.TV_SHOW_KEY, item)
+        startActivity(openIntent)
     }
 }
